+++ Skip.env
@@ -14,7 +14,7 @@ PRODUCT_BUNDLE_IDENTIFIER = me.fromkk.PhotoExhibition
 MARKETING_VERSION = 1.1.0
 
 // The build number specifying the internal app version
-CURRENT_PROJECT_VERSION = 214
+CURRENT_PROJECT_VERSION = 215
 
 // The package name for the Android entry point, referenced by the AndroidManifest.xml
 ANDROID_PACKAGE_NAME = photo.exhibition
