+++ Sources/PhotoExhibition/Skip/BannerContentView.kt
@@ -1,0 +1,20 @@
+package photo.exhibition
+
+import androidx.compose.runtime.Composable
+import com.google.android.gms.ads.AdRequest
+import com.google.android.gms.ads.AdSize
+import com.google.android.gms.ads.AdView
+import androidx.compose.ui.viewinterop.AndroidView
+
+@Composable
+fun BannerContentView(adUnitId: String) {
+    AndroidView(
+        factory = { context ->
+            AdView(context).apply {
+                setAdSize(AdSize.BANNER)
+                setAdUnitId(adUnitId)
+                loadAd(AdRequest.Builder().build())
+            }
+        }
+    )
+}
