+++ Sources/PhotoExhibition/Skip/skip.yml
@@ -1,3 +1,7 @@
 # Configuration file for https://skip.tools project
 build:
   contents:
+    # AdMobの依存関係を追加
+    - block: 'dependencies'
+      contents:
+        - 'implementation("com.google.android.gms:play-services-ads:22.6.0")'
