+++ Versions.env
@@ -2,4 +2,4 @@
 MARKETING_VERSION = 1.1.0
 
 // The build number specifying the internal app version
-CURRENT_PROJECT_VERSION = 214
+CURRENT_PROJECT_VERSION = 215
